cleaned up code using coderush feature.  Cleaned emptystring with string.empty, removed unused variables and private types.  Formatted strings are now string interpolations (aids in readability)

+++ MonoGame.Framework/Audio/AudioUtil.cs
@@ -13,28 +13,28 @@ namespace Microsoft.Xna.Framework.Audio
             //buffer should contain 16-bit PCM wave data
             short bitsPerSample = 16;
 
-            using (var mStream = new MemoryStream(44+buffer.Length))
+            using (var mStream = new MemoryStream(44 + buffer.Length))
             using (var writer = new BinaryWriter(mStream))
             {
                 writer.Write("RIFF".ToCharArray()); //chunk id
-                writer.Write((int)(36 + buffer.Length)); //chunk size
+                writer.Write((36 + buffer.Length)); //chunk size
                 writer.Write("WAVE".ToCharArray()); //RIFF type
 
                 writer.Write("fmt ".ToCharArray()); //chunk id
-                writer.Write((int)16); //format header size
+                writer.Write(16); //format header size
                 writer.Write((short)1); //format (PCM)
                 writer.Write((short)channels);
-                writer.Write((int)sampleRate);
-                short blockAlign = (short)((bitsPerSample / 8) * (int)channels);
-                writer.Write((int)(sampleRate * blockAlign)); //byte rate
-                writer.Write((short)blockAlign);
-                writer.Write((short)bitsPerSample);
+                writer.Write(sampleRate);
+                short blockAlign = (short)((bitsPerSample / 8) * channels);
+                writer.Write((sampleRate * blockAlign)); //byte rate
+                writer.Write(blockAlign);
+                writer.Write(bitsPerSample);
 
                 writer.Write("data".ToCharArray()); //chunk id
-                writer.Write((int)buffer.Length); //data size
+                writer.Write(buffer.Length); //data size
 
                 writer.Write(buffer);
-                
+
                 return mStream.ToArray();
             }
         }
